Custom ping-lite module
The original ping-lite module is broken (at least in Windows). It'll always return null just because of a wrong regmatch.

As a result, the whole program is rendered useless, since it rely heavily on the ping result.

Can't check for the other OS, as they should be dependent on how the each system ping output is formatted.